added user attribute to all entities and implemented user registration function

+++ src/main/java/IvanovVasil/OrderManagmentSystem/ingredient/IngredientsService.java
@@ -1,7 +1,9 @@
 package IvanovVasil.OrderManagmentSystem.ingredient;
 
+import IvanovVasil.OrderManagmentSystem.User.User;
 import IvanovVasil.OrderManagmentSystem.exceptions.BadRequestException;
 import IvanovVasil.OrderManagmentSystem.exceptions.NotFoundException;
+import IvanovVasil.OrderManagmentSystem.exceptions.UnauthorizedException;
 import IvanovVasil.OrderManagmentSystem.webSocket.ChatMessageService;
 import IvanovVasil.OrderManagmentSystem.webSocket.ElementToUp;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,58 +20,63 @@ public class IngredientsService {
   @Autowired
   ChatMessageService cms;
 
-  public List<Ingredient> getAllIngredients() {
-    return ir.findAll();
+  public List<Ingredient> getAllIngredients(User user) {
+    return ir.findAllByUserId(user.getId());
   }
 
   public Ingredient save(Ingredient ingredient) {
     return ir.save(ingredient);
   }
 
-  public Ingredient save(IngredientDTO ingredient) {
+  public Ingredient save(IngredientDTO ingredient, User user) {
     Ingredient ingredientFound = ir.findByNameIgnoreCase(ingredient.ingredientName());
-
-    if (ingredientFound != null) {
-      throw new BadRequestException("Ingredient already exists");
-    }
-
     Ingredient newIngredient = Ingredient
             .builder()
             .name(ingredient.ingredientName())
             .ingredientCategory(IngredientCategory.valueOf(ingredient.ingredientCategory()))
+            .user(user)
             .build();
     ir.save(newIngredient);
     cms.sendUpdateMessage(ElementToUp.INGREDIENT);
     return newIngredient;
-
   }
 
   public Ingredient findById(UUID ingredientId) {
     return ir.findById(ingredientId).orElseThrow(() -> new NotFoundException(ingredientId));
   }
 
-  public Ingredient editIngredient(UUID ingredientId, IngredientDTO ingredient) {
+  public Ingredient editIngredient(UUID ingredientId, IngredientDTO ingredient, User user) {
     Ingredient ingredientFound = this.findById(ingredientId);
+    if (ingredientFound.getUser().getId() == user.getId()) {
+      if (!ingredient.ingredientName().isEmpty()) ingredientFound.setName(ingredient.ingredientName());
+      ingredientFound.setIngredientCategory(IngredientCategory.valueOf(ingredient.ingredientCategory()));
+      ir.save(ingredientFound);
 
-    if (!ingredient.ingredientName().isEmpty()) ingredientFound.setName(ingredient.ingredientName());
-    ingredientFound.setIngredientCategory(IngredientCategory.valueOf(ingredient.ingredientCategory()));
-    ir.save(ingredientFound);
-
-    cms.sendUpdateMessage(ElementToUp.INGREDIENT);
-    return ingredientFound;
-  }
-
-  public void delete(UUID ingredientID) {
-    Ingredient found = this.findById(ingredientID);
-    if (ingredientID != null) {
-      ir.deleteById(ingredientID);
       cms.sendUpdateMessage(ElementToUp.INGREDIENT);
+      return ingredientFound;
     } else {
-      throw new BadRequestException("Ingredient id cannot be null");
+      throw new UnauthorizedException("You have not permissions to edit this item!");
     }
+
   }
 
-  public List<Ingredient> getAllIngredientsByCategory() {
+  public void delete(UUID ingredientID, User user) {
+    Ingredient found = this.findById(ingredientID);
+
+    if (found.getUser().getId() == user.getId()) {
+      if (ingredientID != null) {
+        ir.deleteById(ingredientID);
+        cms.sendUpdateMessage(ElementToUp.INGREDIENT);
+      } else {
+        throw new BadRequestException("Ingredient id cannot be null");
+      }
+    } else {
+      throw new UnauthorizedException("You have not the permissions to delete this ingredient!");
+    }
+
+  }
+
+  public List<Ingredient> getAllIngredientsByCategory(User user) {
     return null;
   }
 }
